+++ src/controller_rpc.h
@@ -38,11 +38,6 @@
 #define _CONTROLLER_RPC_H
 
 /*
- * Constants
- */
-#define CONTROLLER_DEVICE_TIMEOUT_DEFAULT 30
-
-/*
  * Prototypes
  */
 #ifdef __cplusplus