+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## 1.7.0
 Planned: February 2026
 
+### New features
+
+* New `clixon-controller@2025-12-01.yang` revision
+  * Added `service-timeout`
+
 ### Corrected Bugs
 
 * Fixed: [Candidate datastore lock prevents sequential commit operations via RESTCONF](https://github.com/clicon/clixon-controller/issues/236)